Add home-, login-bar- and login-component
The default scaffolding is commented-out, but still accessible. It has 
been replaced with a custom, but not yet functional UI-Mockup. It 
consists out of a home screen and a login screen, which is still 
default.

Also navigation has been added to the application. The 
Angluar-RoutingModule is used for this purpose.