Add login-, create-account- and input-component
The input-component is a simple component, which makes it a little 
easier to create multiple inputs.

The login- and create-account-component do not have any functionality at 
the moment. These components are a mockup of the UI. You can navigate 
inbetween all the screens which include the login and create-account 
screen.