Implement functionality to login- and create-user-component
The login- and create-user-component now have the functionality to store 
the inputs. The inputs get stored inside a model, which is specific for 
each component. There are now three new input-types wich all work with 
'ngModel', and are therefore easy to work with. The same applies for the 
select-component.

Further there are additional classes for later use (input validation).


If you want to use the functionality: The inputs get printed to the 
console after clicking the 'Next' button.