Add httpClient to login- and create-account-component
The both components are now functional, which means that you can now 
register and log in. The logout method has not been created yet. Both 
components now send requests to the backend, which handles the rest. The 
methods in the backend were already implemented. The only thing that has 
been changed in the backend is the model for the user, which now has all 
the necessary attributes the user needs. 

The input handeling is also simpler now, which is why the two models in 
the frontend are gone now.

The login-bar-component now shows if a user is loggend in and with what 
username.