Add mockup buy, post and approve button
The homepage now has a buy, post and approve button. The buy and post button are mockups of the actual 'buy a product' and 'post a product' process. When clicked each of these buttons will give you a message displayed on the screen. The approve button is not yet implemented, because one would need to have admin users.