Add custom validators for input validation
Since there are custom inputs which use the ControlValueAccessor interface, one can define custom validators which handle the validation of the inputs. The validators are modular, which means that they can be applied to any input which uses the interface above. 

The new validators all use one base class: RegexValidatorBase. This class allows for easy definition of validators with any given regex. The predefined regex's can be found in the base class file. 

Right now there are two validators: 
- PasswordValidatorDirective and
- EmailValidatorDirective
They are already applied in the LoginComponent. 

Other changes are rearrangements in the directory structure.