Input validation for create-account and login active
The input validation for the login and the registration are now active. This means that the validator directives for every implemented input exists. You can find the validators under the validator directory in the frontend.

Right now the input does not get accepted if something is wrong, but there is no indication to what is wrong.