Display validation error-messages better
The error-messages from the input validation are now handled with a component which you pass on the form from which you woud like the messages to be handled.

When passed the form the error-messages-component takes the form and iterates over each control of the form. For each control a seperate error-message-component is beeing created. The error-message-component takes the control as an input. From this input one can acces the valueAccessor and the error-messages of the Validators.