Set username max. length to 255 characters and migrate to scss
The input-validation for the username has now been set to a max-length of 255 characters.

The frontend migrated from css to scss and bootstrap is available too. For you to use bootstrap in a scss file, you need to import the root scss file from bootstrap with this line: <<@import "~bootstrap/scss/bootstrap";>> at the top of the file.