Add ProductService and PostProductComponent
The post process for products is now implemented in the frontend.
The ProductService makes use of the PostProductService, which itself uses different models and interfaces. The post method of the ProductService takes a PostProductRequestBuilder as an argument, on which the post method calles the build method. 

To psot a product you have to inject the product service into the component.

Other changes are small cleanups in the backend.