Add navigation to user profile
The UserProfileComponent and AdminProfileComponent have been removed and replaced with a single ProfileComponent. This component has a navigation with three points: 'User Details', 'My Products' and 'Create New Product'. These points are rendered using the ProfileNavigationComponent. This component then uses the Router module, to route the different buttons/points to its corresponding URLs. The URLs can be found in app.module.ts. Each URL then has a defined component to render. For this to work there is need for two new components: MyProductsComponent and UserDetailsComponent (at the moment default components).