Implement extension for admin profile and style profile
The profile is now extendable for the admin profile. It is not implemented yet but it's simpler now: The ProfileComponent now works with the Router Module of Angular. It receives data, which default Component to render. This then is extendable for different routes with different data, for example: 'user/admin/profile/details' can receive other data than 'user/profile/details', but they render the same component. The only concern is security. Right now everyone can access the profile pages.

The profile page now also looks decent.