Update UserService with more functionality
The UserService does now store the user on login. The information then also gets stored in the localStorage. You can now use the service to check if a user is an admin or if he is logged in. You can also retreive data from the service, but you cannot change it.