Add get request for user with userId
In the backend, there is now a secure endpoint called 'user/userid::userId'. This endpoint is used for get requests. If ther request is successfull, the response body contains all user information from the database. The information then gets stored in the UserService. The user services login function also returns a promise now instead of a observable. The Promise returns a boolean true, if the login request was ok and the error message if not. 

The menubar uses now the UserService instead of the localStorage, as well as the ProfileComponent.