Add ApproveProductComponent and async pipe for profile user information
The ApproveProductComponent is just a default component at the moment.

On the ProfileComponent the get-request from the UserService on reload takes some time, therefore the infromation is not present when the frontend has finished loading. The field is at this point in time still null and an error occurs. The solution is to use async pipe and observables for the variables.