Fix Bug of asynchronous requests with HttpClient
The problem was that the frontend made requests to the backend and needed an answer in order to display information. This would easily work with a subscription. But since the frontend made multiple subscribtions to the Observable the request returned, the frontend made multiple requests to the backend. This is enormously time inefficient and uses way to many resources of the backend.

The soulution to this problem was to pipe the observable with the share operator. This operator tells the observable to use the same data in all the subscriptions. Therefore there are no additional requests to the backend.

This method is now applied to all the user services which make a request. Also this get request is now its own service.

The components also have been updated correctly, according to the changes (async pipe has been removed).