Add AuthGuard
The AuthGuard is now implemented and used by the RouterModule. It uses the UserService to check if a user is logged in or not. If not, the guard denies access to the 'user/profile' url and redirects the user to 'user/login'. If the user is logged in, the guard provides access to the above url.