Implement ProductCardGridComponent and add ProductPillCompnent
The grid card is now implemented. This component is used in the approve-products component. The approve component uses the card if a boolean (displayList) is set to false, otherwise it will use the list-card. The approve comonent is therefore able to switch between views, if only hard-coded. This can be changed in the future.

The cards, as well as the information component use the new pill component, which lets you create pills with lots of customization. You can control szie, color and value of the pill by passing methods and strings, etc.