Implement checking for creator equals current user
The product-information component checks if the current user is the creator of the product displayed. If this is the case the options for interaction with the product change accordingly.

After the component gets the product, and therefore the userId of the creator, the component gets the creator. This needed some backend implementation. A middleware has been created, so that if a token is supplied, the whole user gets returned (if token is verifyed), else the information about the user id decimated to userName, email and userId. 

After the creator is retrieved, the current is queried, if the user is logged in. If not there will be a null user. The Component then makes a comparison of the two userIds of the creator and the currentUser. This comparison then tells the component which options to display.

Also the missing information has been added to the component.