Implement input validation for post product
The input now gets validated in the post product component. There is a new category of validators for selects. The validators use predefined options to build a regexp. It then determines wether the input is valid or not using the regexp.

There is also a new custom input (textarea). The textarea is just like any other normal custom input and takes the same inputs. Additionally it takes a 'rows' input. 

A lot of files have been modified with minor changes.