Split product information and add guards
The product information component had to handle too many things at the same time and needed to be easily extendable, which it was not. Therefore the component was split up into smaller components, which all can be used individually. The advantage of this is, that the preview is fixed and each page has its own url aswell as its own guard. This means you cannot access the purchase page if you are the creator. You can only edit. The same applies for the other way around. If you are not logged in, you can access the purchase page, but no other. As an admin you can access the review page of a product.