Add basic logic for reject and accept buttons
The accept and reject buttons now both show another form/button when clicked. This form/button is there to prevent accidential clicks. The admin has to cklick twice to accept a product as well as enter a reason of rejection to reject the product.

The styling of the added buttons and forms is not very good at the moment.