Add rejected products component on profile
The rejected products component is a default component which should later display all the rejected products. 

For easier implementation of further components on the profile page, the profile navigation elements now contain a component, which gets diplayed on the specified route. The whole array can then be passed on to the router config.