Add frontend implementation to get all rejected products
All rejected products can now be retreived using the ProductService. The method is called getMyRejectedProducts in both Product- and GetProductService. A backend bugfix was also made. 

All the rejected products get now displayed on the rejected products page on the profile page. If a product is rejected, it will show the rejection message on the edit product page.