Add possibiliy to switch theme from bright to dark
The user profile now has a navigation element called preferences. In these preferences a single checkbox can be found and clicked. When clicked the theme changes from bright to dark and vice-versa. 

For this to work there is a new Service called ThemeService. The service holds all possible themes and can switch inbetween them. Then there is a Themable class which components can extend. This class adds itself to the ThemeServices themable array. When the switchTheme method gets called, all the themables get notifyed that the theme has changed.