Complete rewrite of validation
The validation process is completely rewritten. As a user you now receive multiple error messages to what is wrong with the input.

The API is very simple and intuitive. You can choose between constraints. They are based on simple things like 'start with a capital letter' etc. You can add the different constraints to an array and combine the arrays in logical ORs or ANDs. There is also an operator class which lets you define new operators if you wish. The selection of operators is still hard coded at the moment.