Add custom datetime-locale input
The date input takes the same inputs as a normal input and addtional optional min and max. The min and max let you define, which dates can be chosen from min-date to max-date. 

The scss for this component is similar to a normal input, but has some minor changes to the label.