Implement dynamic stages creation for stage inidcator
The stage indicator needs to have titles for the different stages. Therefore it is practical to define the stages all in one place. This can only be done by using the ViewContainerRef and dynamically insert the components.

So the BuyItemComponent now defines an array of StageModels, which contains the different stages and titles of the stages.