Implement navigation and data emitting
The navigation of a stage is complicated. The navigation uses eventemitters in the single stages. The stages then publish the click on 'Next' to the Stagable, which handles all the navigation logic. Also the data can be retreived from a Stage using the dataEmitter field on the Stage abstract class.