Redesign of Stages
The stages needed a complete redesign of everything, because there was an unsolvable problem with the existing design.

The new design uses directives: StageNavEmitter and StageNavDataEmitter<T>. These directives are extended by components which should be used as stages. These components then use the stage component as a base element in their html. Into the stage component you can pass on the naviation and the content for this specific stage. This makes it very easy and expandable. 

It is necessary, that the components which represent a stage implement the StageNavigationDataEmitter directive, because you pass these stages to the Stagable directive, which needs the directive.

The Stagable directive then takes each stage and builds the corresponding component and inserts it into the DOM. The directive also assigns the EventEmitters, defined in the NavData directive, to specific methods: Set Data, Change Stages.