Add theme directive
The theme directive is supposed to replace the Themable class. The directive is a much more elegant and simpler way to apply theming to the website. 

The theme directive can be used in combination with the button OR buttonEmpty directive, not both. You can also manually assign theme extentions, which will appear in the class of the element, for example: [classExtentions]="['hello', 'bye']" will add the following classes:
"hello-${theme} bye-${theme}"