Add seperate classes for map
There is a Map, Locations and Search component, which can be implemented by a component. These classes are used to simplify the usage of the map on a component. You simply pass on the ElementRef for the map, the baseLayer and some options for initial position and zoom and you have a map in the DOM. Either with a search bar or just one which can show locations.