Implement input logic and inputs
The logic and the instructions for the address input is now implemented. The instructions tell you how to enter the address with the help of the map. The map displays the users home address by default. If a user wishes to ship the item to his home he just has to approve the home address and the next button becomes available. If he wishes to ship the item to a different address, the can enter the address on the map and can then click the next button.