Implement logic for payment method selection
The payment method can now be selected and the data is stored in the paymentMethod component. The data gets emitted, if the stages are finalized. This means that the payment method component overrides the finalizeStages method of the StagesNDEExtention and emits the data in this method. There is also a basic implementation for the handling of the data handling on the buy item component.