Add order service and all types
The order service has a putOrder method of the PutOrderService which sends a request to the backend at a specific endpoint url. This url is specified in the stagable components, here BuyItemComponent. This class also implements a request getter and a getter for the url. The request getter method needs a buildOrderRequest method, which then returns the request for the order. The order request is typed. The types can be found in the requests folder as well as the responses (in responses folder).