Create Preferences and associate tables
The preferences table now exists with its properties. The table has a foreign key to the users table on the userId. OnDelete and Update it will cascade. Also the preferencesController does exist and work. One last thing is that on creation of a user, the preferences table is going to create an entry automatically.