Improve httpClient process
The preference service is the first prototype of this method and it works pretty well. The goal is, that you can add different subscribers for different events. These events concerne a particular service. For example the product can be updated, therefore one could use the onUpdate event to do stuff on the update. This pattern can be used for everyting. Right now there are two predefined events: OnLoad and OnUpdate.