Add address table with associations
The new address table has an association to the users table: One user can have exactly one address, but one address can belong to multiple users (one to many). Therefore the register process now checks if the address already exists and only inserts the new addres, if and only if it does not exist. The new user gets inserted eitherway.