Bug Fix on user details
The user details map pushed an address string to the map component. The problem was that not the address itself was pushed but the string of the address. This lead to a toString call of the undefined address. This is now fixed by pushing the address object to the map.