Add map to register component
The register component now has a map for the address of the product. For the address to work, every model in the frontend had to change and update. Also the searching component needed to be changed. The searching component now uses the product service to search for products. The search is implemented with a filter called Search. This filter can test the products. This can then be used in combination with the array.filter method.