Update Models and Implement createItemSold with transaction
The Models have been updated. This means, that all the associations have been reworked and clearly stated. The backend now needs testing.

The createItemSold method has been implemented with transactions. The method gets the shipping address form the database or creates a new one. Then it creates an order and an ItemSold.