Change update process of purchase products
The problem was that the transaction blocked all access to modelinstances wich were created in the same transaction. This means that when getting the freshly cerated order of the suborder, the order is not found and therefore its products neither. The workaround is to check (validation) and get all the necessary instances beforehand and modify them inside the transaction.

This solution is implemented at the moment.