Improve User Service as well as Theme Service
The User Service now uses new methods for different events for example there is a event for when the UserService loaded a user into its storage. To this event can be subscribed via a onLoad method on the service.
The implementation is based on a way that allows you to easily define custom events. The restriction is that the events have to do with observables. But it can be any observable. 

The theme service had a bug which is now fixed. The Themable class is now completely removed from the code. The class has been replaced with a ThemeObserver Interface.