Update User service for better events
The events on the userservice are now distributed differently. This makes the process of loading a user much simpler and faster. The userservice now only loads the user from the login service. The user service does not now anything about the login process and does not have to call login methods. All the userservice does is load the user from the login service on login.