Fix Services and improve design
The services now use a new design following the observer pattern with an extension. The extension is loading. A service can now load a source and then notify the observer when starting, loading and finishing. The observer then also receives either an error or the value loaded. 

The other fixes were fixes with scss implementations.