Improve Card Components
The cards components like the login, register etc are now setup in a more generic way, which allows the components to be used in a non centered way, for example on the user profile. This is not implemented yet.

The implementation is done with a card component (CenterCardComponent), which just takes ng content, title and error message as an input.