Fix Bug in update process
The infromation was correctly sent to the backend but the problem was twofold: The file was not defined, because multer could not parse the image if no picture was sent and the second one was that the address was JSON stringified in the frontend but not parsed in the backend. Therefore the DB could not insert the address, becaues it was a string. See in ProductController