Implement messages on product details
The product details has now a messanger which creates a new thread if there is no existing one. The problem at the moment is, that every user can now send a message, which means, that the seller can send a message to himself.

The message service is implemented and uses basic data from the backend. Only the creation of a trhead needs to be tested and improved.