Fix thread accepting bug and message display
The backend check, if the users who sent the messages belong to the tread, had errors in it but is now fixed and moved to the thread model. 

I the frontend the participants werde stored wrong and also checked wrong, threrfore the participants did not get dispalyer right. The sending of a message also did result in a bad request, because the sender was not correct.

The message service sets now the current sender of the messaged directly and not the components, whch is much more efficient and cleaner.