Implement update of read status of message
The message read status now gets updated whenever a user selects a thread. On select every message which is not from the user, will be updated and set to read. 

The backend implmenetation was simply creating another controller entry and update a threads messages read status. In the frontend the message service had to be updated and also the merging of the threads had to be updated.